Refactor common code from PeoplePage and ProjectsPage
model.js was basically the same for both
