+++ src/components/pages/PeoplePage/model.js
@@ -15,66 +15,11 @@
  */
 import {Rx} from '@cycle/core';
 import _ from 'lodash';
-import moment from 'moment';
-import {smartStateFold} from 'power-ui/utils';
+import dtModel from 'power-ui/components/pages/common/data-table-page-model';
 
-const initialState = {
-  people: [],
-  progress: 0,
-  filtered: [],
-  tribes: [],
-  timeRange: { // selected time range
-    start: moment().startOf('month'),
-    end: moment().clone().add(2, 'months').endOf('month'),
-  },
-  availableTimeRange: {
-    start: moment().startOf('month'),
-    end: moment().clone().add(5, 'months').endOf('month'),
-  },
-  filters: {
-    location: 'all',
-    search: '',
-    availability: null,
-  },
-};
-
-function makeUpdateFn$(peopleData$, props$, actions) {
-  const updatePeopleArray$ = peopleData$
-    .map(({dataArray, progress}) => function updateStateWithPeopleArray(oldState) {
-      return {...oldState, people: dataArray, progress};
-    });
-
-  const updateStateWithProps$ = props$
-    .map(props => function updateStateWithProps(oldState) {
-      return {
-        ...oldState,
-        tribes: props.tribes || oldState.tribes,
-        availableTimeRange: props.availableTimeRange || oldState.availableTimeRange,
-        filters: {
-          ...oldState.filters,
-          location: props.location || oldState.filters.location,
-        },
-      };
-    });
-
-  const updateSearchFilter$ = actions.setSearchFilter$
-    .map(search => function updateStateWithSearchFilter(oldState) {
-      return {...oldState, filters: {...oldState.filters, search}};
-    });
-
-  return Rx.Observable.merge(
-    updatePeopleArray$,
-    updateStateWithProps$,
-    updateSearchFilter$
-  );
-}
-
-function model(peopleData$, props$, actions) {
-  const update$ = makeUpdateFn$(peopleData$, props$, actions);
-  const state$ = update$
-    .scan(smartStateFold, initialState)
-    .shareReplay(1);
-  return state$;
+function model(...args) {
+  return dtModel(...args)
+    .map(state => ({...state, people: state.dataArray}));
 }
 
 /**
